+++ event/type.go
@@ -187,7 +187,9 @@ var (
 	StateHalfShotBridge    = Type{"uk.half-shot.bridge", StateEventType}
 	StateSpaceChild        = Type{"m.space.child", StateEventType}
 	StateSpaceParent       = Type{"m.space.parent", StateEventType}
-	StateInsertionMarker   = Type{"org.matrix.msc2716.marker", StateEventType}
+
+	// Deprecated: MSC2716 has been abandoned
+	StateInsertionMarker = Type{"org.matrix.msc2716.marker", StateEventType}
 )
 
 // Message events